// SPA/src/app/Model/IIntroductionUpdate.ts
export interface IIntroductionUpdate{
  
    id: string;
    introductionRequestId: string;
    userFromId: string;
    userFromName: string;
    userMiddleId: string;
    userMiddleName: string;
    userToId: string;
    userToName: string;
    presentationMessage: string;
    introductionMessage: string;
    introductionState : string;
    active: string;

  }